pyproject.toml
[project]
name = "mybazar-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
]
